emplyees RUD functionalities finished

+++ web2-app/src/app/employees/employees.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, SystemJsNgModuleLoaderConfig } from '@angular/core';
 import {Employee} from '../employees';
 import { EmployeeService } from '../employee.service';
+import { DepartmentService } from '../department.service';
+import { Department } from '../departments';
 
 
 @Component({
@@ -11,6 +13,7 @@ import { EmployeeService } from '../employee.service';
 export class EmployeesComponent implements OnInit {
 
   employees:Employee[];
+  departments: Department[];
 
   constructor(private employeeService: EmployeeService) { }
 
@@ -23,18 +26,24 @@ export class EmployeesComponent implements OnInit {
     .subscribe(employees => this.employees = employees);
   }
 
-  add(first_name: string): void {
+  add(first_name: string, last_name: string): void {
     first_name = first_name.trim();
+    last_name = last_name.trim();
+    //birth_date = null;
+    //dep_number = this.departments[0].id;
+
     if (!first_name) { alert("Please type a name"); return; }
-    this.employeeService.addEmployee({ first_name } as Employee)
+    this.employeeService.addEmployee({ first_name, last_name} as Employee)
       .subscribe(employee => {
         this.employees.push(employee);
       });
   }
 
   delete(employee: Employee): void {
+    if(confirm("Are you sure?")) {
     this.employees = this.employees.filter(h => h !== employee);
     this.employeeService.deleteEmployee(employee).subscribe();
+    }
   }
